// frontend/src/types/api/listCalendarEventsApiCalendarEventsGetSort.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * mailassist
 * OpenAPI spec version: 0.1.0
 */

export type ListCalendarEventsApiCalendarEventsGetSort = typeof ListCalendarEventsApiCalendarEventsGetSort[keyof typeof ListCalendarEventsApiCalendarEventsGetSort];


export const ListCalendarEventsApiCalendarEventsGetSort = {
  newest: 'newest',
  oldest: 'oldest',
  title: 'title',
} as const;